Backend/requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cors==0.9.7
cryptography==3.0
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.6.0
django-heroku==0.3.1
django-nucleus==0.1.2
djangorestframework==3.11.0
Faker==4.0.0
filelock==3.0.12
Flask==1.1.1
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
gevent==20.12.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
logzero==1.5.0
lyricwikia==0.1.11
MarkupSafe==1.1.1
mccabe==0.6.1
mutagen==1.43.0
pafy==0.5.5
paramiko==2.7.1
pathlib==1.0.1
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
pybin==0.3.4
pycparser==2.19
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
PyNaCl==1.4.0
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-file==1.5.1
six==1.14.0
soupsieve==1.9.5
spotdl==1.2.4
spotipy==2.6.3
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
titlecase==0.12.0
tldextract==3.1.0
typed-ast==1.4.1
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.11.2
youtube-dl==2020.1.24
zope.event==4.5.0
zope.interface==5.2.0
